% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnormfluodbf_bar.R
\name{barplots}
\alias{barplots}
\alias{ggnormfluodbf_vbar}
\alias{ggnormfluodbf_vbar_alt}
\alias{ggnormfluodbf_hbar}
\title{Bar plots generic}
\usage{
ggnormfluodbf_vbar(data, mapping, size, v_adj = 0.5, fill_colors = NULL, ...)

ggnormfluodbf_vbar_alt(data, x, y, size, fill_colors = NULL, ...)

ggnormfluodbf_hbar(data, mapping, size, h_adj = 0.5, fill_colors = NULL, ...)
}
\arguments{
\item{data}{data}

\item{mapping}{aesthetic vector}

\item{size}{text size}

\item{v_adj}{vertical adjustment}

\item{fill_colors}{fill colors}

\item{...}{dots}

\item{x}{x}

\item{y}{y}

\item{h_adj}{horizontal adjustment}
}
\value{
ggplot object

ggplot object

ggplot object

ggplot object
}
\description{
Bar plots generic
}
\examples{
\dontrun{
ggnormfluodbf_vbar(tips, mapping = aes(x = smoker, y = sex), size = 8, fill_colors = c('red','blue'))
ggnormfluodbf_hbar(tips, mapping = aes(x = sex, y = smoker), size = 8, fill_colors = c('maroon','gray'))
ggnormfluodbf_hbar(df1, mapping = aes(x = sex, y = smoker), size = 8, fill_colors = c('maroon','blue','green'))
ggnormfluodbf_hbar(df2, mapping = aes(x = sex, y = smoker), size = 5, fill_colors = c('red','blue','maroon','green'),h_adj = 0.4))
ggnormfluodbf_hbar(df2, mapping = aes(x = smoker, y = sex), size = 5, fill_colors = c('red','blue','maroon','green'),h_adj = 0.4)}
}
