% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_functions_dictionary}
\alias{create_functions_dictionary}
\title{Create dictionary in .dcf format}
\usage{
create_functions_dictionary()
}
\arguments{
\item{key}{key}

\item{func}{value}

\item{filename_with_dcf_ext}{file name with dcf extension}

\item{temp_loc}{temp boolean}
}
\description{
Create dictionary in .dcf format
}
\examples{
\dontrun {
tp <- function(package = "./", ...){
devtools::document(package)
devtools::test(package, ...)}
key <- "test_package"
dict <- create_functions_dictionary()
file_path <- dict$add_field(field_name = "bar_data", mbd, "hf.dcf")}
}
