% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnormfluodbf-package.R
\docType{package}
\name{ggnormfluodbf-package}
\alias{ggnormfluodbf}
\alias{ggnormfluodbf-package}
\title{ggnormfluodbf: A package designed to plot data from normfluodbf and to immerse myself into the grammar of R graphics}
\description{
A graphics package designed to support the R package Normfluodbf, authored by Tingwei and Tesla Adeck.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlphaPrime7/ggnormfluodbf}
  \item \url{https://alphaprime7.github.io/ggnormfluodbf/}
  \item Report bugs at \url{https://github.com/AlphaPrime7/ggnormfluodbf/issues}
}

}
\keyword{internal}
