% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggnormfluodbf_aes}
\alias{ggnormfluodbf_aes}
\title{ggnormfluodbf aesthetic mapper}
\usage{
ggnormfluodbf_aes(...)
}
\arguments{
\item{...}{dots}
}
\value{
aes list
}
\description{
ggnormfluodbf aesthetic mapper
}
\examples{
\dontrun{
ggnormfluodbf_aes(x, y)
}
}
