% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnormfluodbf_box.R
\name{ggnormfluodbf_box_labelgeom_deprecated}
\alias{ggnormfluodbf_box_labelgeom_deprecated}
\title{Ggnormfluodbf Box Plot}
\usage{
ggnormfluodbf_box_labelgeom_deprecated(
  data,
  mapping,
  alpha = 0.5,
  dodge_pos = 0.75,
  vadj = 1.5,
  hadj = 0,
  include_mean = TRUE,
  include_labels = TRUE,
  label_size = 3.5,
  label_color = "red",
  label_alpha = 0.8,
  label_family = "serif",
  label_fontface = "bold",
  label_box_padding = ggplot2::unit(0.5, "lines"),
  label_box_alpha = 0.8
)
}
\arguments{
\item{data}{data}

\item{mapping}{mapping}

\item{alpha}{alpha}

\item{dodge_pos}{dodge_pos}

\item{vadj}{vertical adjustment}

\item{hadj}{horizontal adjustment}

\item{include_mean}{mean boolean}

\item{include_labels}{labels boolean}

\item{label_size}{label size}

\item{label_color}{label text color}

\item{label_alpha}{color intensity}

\item{label_family}{font family}

\item{label_fontface}{font face}

\item{label_box_padding}{padding}

\item{label_box_alpha}{label box color intensity}
}
\value{
ggplot object
}
\description{
Ggnormfluodbf Box Plot
}
\examples{
\dontrun{
ggnormfluodbf_box_labelgeom_deprecated(tips,ggplot2::aes(x = smoker, y = tip, fill = sex))
ggnormfluodbf_box_labelgeom_deprecated(tips,ggplot2::aes(x=smoker, y = tip, fill = sex), alpha = 2)
ggnormfluodbf_box_labelgeom_deprecated(tips,ggplot2::aes(x=smoker, y = tip, fill = smoker), alpha = 2, vadj = -5, hadj = 0.7, include_labels = F, include_mean = F)
ggnormfluodbf_box_labelgeom_deprecated(tips,ggplot2::aes(x=smoker, y = tip, fill = sex), alpha = 2, vadj = -5, hadj = 0.7)
ggnormfluodbf_box_labelgeom_deprecated(tips,ggplot2::aes(x=smoker, y = tip, fill = sex), alpha = 2, vadj = -5, hadj = 0.7, include_labels = F)
ggnormfluodbf_box_labelgeom_deprecated(tips,ggplot2::aes(x=smoker, y = tip, fill = sex), alpha = 2, vadj = -5, hadj = 0.7, include_labels = F, include_mean = F)
}
}
