% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnormfluodbf_box.R
\name{ggnormfluodbf_grouped_box}
\alias{ggnormfluodbf_grouped_box}
\title{Ggnormfluodbf Box Plot}
\usage{
ggnormfluodbf_grouped_box(
  data,
  mapping,
  alpha = 0.5,
  dodge_pos = 0.75,
  vadj = -0.5,
  hadj = 0.5,
  include_mean = TRUE,
  include_labels = TRUE
)
}
\arguments{
\item{data}{data}

\item{mapping}{mapping}

\item{alpha}{alpha}

\item{dodge_pos}{dodge_pos}

\item{vadj}{vertical adjustment}

\item{hadj}{horizontal adjustment}

\item{include_mean}{mean boolean}

\item{include_labels}{labels boolean}
}
\value{
ggplot object
}
\description{
Ggnormfluodbf Box Plot
}
\examples{
\dontrun{
ggnormfluodbf_grouped_box(tips,ggplot2::aes(x = smoker, y = tip, fill = sex))
ggnormfluodbf_grouped_box(tips,ggplot2::aes(x=smoker, y = tip, fill = sex), alpha = 2)
ggnormfluodbf_grouped_box(tips,ggplot2::aes(x=smoker, y = tip, fill = sex), alpha = 2)
ggnormfluodbf_grouped_box(tips,ggplot2::aes(x=smoker, y = tip, fill = sex), alpha = 2, vadj = -5, hadj = 0.7)
ggnormfluodbf_grouped_box(tips,ggplot2::aes(x=smoker, y = tip, fill = sex), alpha = 2, vadj = -5, hadj = 0.7, include_labels = F)
ggnormfluodbf_grouped_box(tips,ggplot2::aes(x=smoker, y = tip, fill = sex), alpha = 2, vadj = -5, hadj = 0.7, include_labels = F, include_mean = F)
}
}
