% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnormfluodbf_pie.R
\name{ggnormfluodbf_pie}
\alias{ggnormfluodbf_pie}
\title{Ggnormfluodbf Pie Charts}
\usage{
ggnormfluodbf_pie(data, mapping)
}
\arguments{
\item{data}{data}

\item{mapping}{mapping}
}
\value{
ggplot object
}
\description{
Ggnormfluodbf Pie Charts
}
\examples{
\dontrun{
ggnormfluodbf_pie(df1, ggplot2::aes(x=smoker))
ggnormfluodbf_pie(df1, ggplot2::aes(x=sex))
}
}
