% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnormfluodbf_scatter.R
\name{ggnormfluodbf_scatter}
\alias{ggnormfluodbf_scatter}
\title{Single Well Scatter Plot}
\usage{
ggnormfluodbf_scatter(
  data,
  mapping,
  title = "Normfluodbf scatter plot",
  xlab = "Cycle Number",
  ylab = "Fluorescence"
)
}
\arguments{
\item{data}{data}

\item{mapping}{aesthetic mapping container}

\item{title}{title}

\item{xlab}{x label}

\item{ylab}{y label}
}
\value{
ggplot object
}
\description{
Single Well Scatter Plot
}
\examples{
\dontrun{
ggnormfluodbf_scatter(dat_1,mapping = aes(x=Cycle_Number, y=A1), colour = 'green')
}
}
