% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_if_interactive}
\alias{print_if_interactive}
\title{Interactive Print}
\usage{
print_if_interactive(p)
}
\arguments{
\item{p}{function}
}
\description{
Interactive Print
}
\examples{
\dontrun{print_if_interactive(p)}
}
