% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_functions_dictionary}
\alias{read_functions_dictionary}
\title{Read a function from the functions dictionary}
\usage{
read_functions_dictionary(key)
}
\arguments{
\item{key}{key}
}
\description{
Read a function from the functions dictionary
}
\examples{
\dontrun{read_functions_dictionary(test_package)}
}
