% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{require_namespaces}
\alias{require_namespaces}
\title{Require Namespace}
\usage{
require_namespaces(pkgs)
}
\arguments{
\item{pkgs}{requisite pkgs}
}
\description{
Require Namespace
}
\examples{
\dontrun{require_namespaces(pkgs)}
}
