% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_invisible_variable}
\alias{set_invisible_variable}
\title{Active Binding in R}
\usage{
set_invisible_variable(name, value)
}
\arguments{
\item{name}{name}

\item{value}{value}
}
\description{
Active Binding in R
}
\examples{
\dontrun{
set_invisible_variable('title',"")}
}
