% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_c}
\alias{str_c}
\title{String Concatenate}
\usage{
str_c(..., sep = "", collapse = NULL)
}
\arguments{
\item{...}{dots}

\item{sep}{separator}

\item{collapse}{collapse}
}
\description{
String Concatenate
}
\examples{
\dontrun{str_c("please install the package '", pkg, "'.  install.packages('", pkg, "') ")}
}
