% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xbarplot_data_summary}
\alias{xbarplot_data_summary}
\title{X Bar Data Summary}
\usage{
xbarplot_data_summary(data, map, h_adj)
}
\arguments{
\item{data}{data}

\item{map}{ggnormfluodbf aesthetic map}

\item{h_adj}{horizontal adjustment}
}
\value{
data summary
}
\description{
X Bar Data Summary
}
\examples{
\dontrun{
weird <- aes(x=smoker,y=sex)
xbarplot_data_summary(data,map = ggnormfluodbf_aes(x=sym(as_name(weird$x)),
y=sym(as_name(weird$y))))
}
}
