% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ybarplot_data_summary}
\alias{ybarplot_data_summary}
\title{Y Bar Data Summary}
\usage{
ybarplot_data_summary(data, map, v_adj)
}
\arguments{
\item{data}{data}

\item{map}{ggnormfluodbf aesthetic map}

\item{v_adj}{vertical adjustment}
}
\value{
data summary
}
\description{
Y Bar Data Summary
}
\examples{
\dontrun{
weird <- ggplot2::aes(x=smoker,y=sex)
ybarplot_data_summary(data, map = ggnormfluodbf_aes(x=sym(as_name(weird$x)),
y=sym(as_name(weird$y))))
}
}
