% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnormfluodbf_bar.R
\name{barplots}
\alias{barplots}
\alias{ggnormfluodbf_vbar}
\alias{ggnormfluodbf_hbar}
\title{Bar plots generic}
\usage{
ggnormfluodbf_vbar(
  data,
  mapping,
  stat = c("prop", "prop_alt", "count"),
  size = NULL,
  v_adj = 0.5,
  fill_colors = NULL,
  bar_args = NULL,
  count_bar_args = NULL,
  ...
)

ggnormfluodbf_hbar(
  data,
  mapping,
  stat = NULL,
  size,
  h_adj = 0.5,
  fill_colors = NULL,
  ...
)
}
\arguments{
\item{data}{data}

\item{mapping}{aesthetic vector}

\item{stat}{stat (prop or count)}

\item{size}{text size}

\item{v_adj}{vertical adjustment}

\item{fill_colors}{fill colors}

\item{bar_args}{bar arguments}

\item{count_bar_args}{count bar arguments}

\item{...}{dots}

\item{h_adj}{horizontal adjustment}
}
\value{
ggplot object

ggplot object

ggplot object
}
\description{
Bar plots generic
}
\examples{
\dontrun{
ggnormfluodbf_vbar(tips, mapping = aes(x = smoker, y = sex), size = 8, fill_colors = c('red','blue'))
ggnormfluodbf_vbar(tips, mapping = aes(x = smoker, y = sex), size = 8, stat = 'prop', fill_colors = c('red','blue'))
ggnormfluodbf_vbar(tips, mapping = aes(x = smoker, y = sex), size = 8, stat = 'count', fill_colors = c('red','blue'))
ggnormfluodbf_vbar(tips, mapping = aes(x = smoker, y = sex), size = 8, stat = 'prop_alt', fill_colors = c('red','blue'))
ggnormfluodbf_hbar(tips, mapping = aes(x = sex, y = smoker), size = 8, fill_colors = c('maroon','gray'))
ggnormfluodbf_hbar(df1, mapping = aes(x = sex, y = smoker), size = 8, fill_colors = c('maroon','blue','green'))
ggnormfluodbf_hbar(df2, mapping = aes(x = sex, y = smoker), size = 5, fill_colors = c('red','blue','maroon','green'),h_adj = 0.4))
ggnormfluodbf_hbar(df2, mapping = aes(x = smoker, y = sex), size = 5, fill_colors = c('red','blue','maroon','green'),h_adj = 0.4)
}
}
