% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnormfluodbf_pie.R
\name{ggnormfluodbf_pie}
\alias{ggnormfluodbf_pie}
\title{Ggnormfluodbf Pie Charts}
\usage{
ggnormfluodbf_pie(
  data,
  mapping,
  label = TRUE,
  fill_colors = NULL,
  fill_brewer_palette = NULL,
  ...
)
}
\arguments{
\item{data}{data}

\item{mapping}{mapping}

\item{label}{label}

\item{fill_colors}{fill_colors}

\item{fill_brewer_palette}{fill_brewer_palette}

\item{...}{dots}
}
\value{
ggplot object
}
\description{
Ggnormfluodbf Pie Charts
}
\examples{
\dontrun{
ggnormfluodbf_pie(df1, ggplot2::aes(x=smoker))
ggnormfluodbf_pie(df1, ggplot2::aes(x=sex))
ggnormfluodbf_pie(df1, ggplot2::aes(x=sex), fill_colors = c('red','green','blue'))
ggnormfluodbf_pie(df1, ggplot2::aes(x=sex), fill_brewer_palette = 'Greens')
}
}
